- Added some defensive code in case no images are found.

+++ main.py
@@ -9,6 +9,9 @@ from scrape import get_random_image
 
 image_file_path = os.path.join('/Library/Desktop Pictures/', IMAGE_FILENAME)
 
+image = get_random_image()
+
 # Update desktop image & restart Dock
-urllib.urlretrieve(get_random_image(), image_file_path)
-call('killall Dock', shell=True)
+if image is not None:
+    urllib.urlretrieve(image, image_file_path)
+    call('killall Dock', shell=True)
